requirements.txt
# --- Core server ---
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1

# --- Data & requests ---
pandas==2.2.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2

# --- NBA data API (compatible with our helpers) ---
nba_api==1.4.1